Add comprehensive code quality improvements: type safety, tests, and extensibility
## Type Safety Fixes
- Fix Event constructor initialization to prevent undefined behavior
- Fix integer overflow in Mode1_DrumMachine note length/flam calculations
- Add MIDI channel validation to MIDIScheduler (1-16 range)

## Test Coverage (47 total tests)
- Add test_pattern.cpp: 10 tests for Pattern class
- Add test_song.cpp: 11 tests for Song class
- Add test_midischeduler.cpp: 12 tests for MIDIScheduler
- Existing tests maintained: test_event.cpp (7), test_track.cpp (7)

## Code Quality Infrastructure
- Add .clang-format: standardized C++ formatting configuration
- Add .cpplint: static analysis configuration for embedded code
- Add Constants.h: centralized configuration and magic number elimination
- Add ModeRegistry.h: improved mode management with RAII and type safety

## Architecture Improvements
- Centralized all magic numbers in Constants.h namespace
- Created extensible ModeRegistry for better mode lifecycle management
- Added validation helpers for MIDI values and BPM
- Improved code documentation and inline comments

## Benefits
- Production-ready code with comprehensive test coverage
- Easy extensibility for future mode development
- Better maintainability with centralized constants
- Type-safe mode management
- Consistent code formatting

See CODE_REVIEW.md for detailed analysis and recommendations.

+++ src/modes/Mode1_DrumMachine.h
@@ -63,12 +63,13 @@ public:
     if (velocity == 0) velocity = 100;
 
     // Map length value (0-127) to note duration (10ms - 2000ms)
-    unsigned long noteLength = 10 + (lengthValue * 1990) / 127;
+    // Cast to unsigned long to prevent overflow before division
+    unsigned long noteLength = 10 + ((unsigned long)lengthValue * 1990) / 127;
 
     // Flam: Send a quieter note slightly before main note (if flam > 0)
     if (flamAmount > 0) {
       // Flam time: 5-50ms delay based on flamAmount
-      unsigned long flamDelay = 5 + (flamAmount * 45) / 127;
+      unsigned long flamDelay = 5 + ((unsigned long)flamAmount * 45) / 127;
       // Flam is quieter (60% of velocity)
       uint8_t flamVelocity = (velocity * 60) / 100;
